teht 4.17 valmis

+++ osa 4/controllers/blogs.js
@@ -43,13 +43,26 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
 	try {
-		const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
+		const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+		if (!decodedToken.id) {
+			return response.status(401).json({ error: 'token invalid or missing' })
+		}
 
-		if (!deletedBlog) {
+		const blog = await Blog.findById(request.params.id)
+
+		if (!blog) {
 			return response.status(404).json({ error: 'Blog not found!' })
 		}
+
+		if (blog.user.toString() !== decodedToken.id.toString()) {
+			return response.status(401).json({ error: 'Only the creator can delete this blog' })
+		}
+
+		await Blog.findByIdAndDelete(request.params.id)
 		response.status(204).end()
+
 	} catch (error) {
+		console.error('Error deleting blog:', error.message)
 		next(error)
 	}
 })
